+++ crates/apvm/src/cmd/link_local.rs
@@ -126,19 +126,7 @@ fn create_bin(node_modules_bin_atlaspack: &Path) -> std::io::Result<()> {
     use std::os::unix::fs::PermissionsExt;
     fs::write(
       node_modules_bin_atlaspack,
-      r#"#!/usr/bin/env node
-
-if (process.env.ATLASPACK_DEV === 'true') {{
-  if (typeof process.env.APVM_ATLASPACK_LOCAL === 'undefined') {{
-    throw new Error('APVM_ATLASPACK_LOCAL must be set when ATLASPACK_DEV is true')
-  }}
-  console.log('USING ATLASPACK SOURCES')
-  require(require('path').join(process.env.APVM_ATLASPACK_LOCAL, 'packages/dev/babel-register'))
-  require('@atlaspack/cli/src/bin.js')
-}} else {{
-  require('@atlaspack/cli/bin/atlaspack.js')
-}}
-"#,
+      "#!/usr/bin/env node\nrequire('@atlaspack/cli/bin/atlaspack.js')\n",
     )?;
     fs::set_permissions(node_modules_bin_atlaspack, Permissions::from_mode(0o777))?;
   }