+++ crates/lmdb-js-lite/__benchmarks__/read-throughput-safe.ts
@@ -5,7 +5,7 @@ const MAX_TIME = 10000;
 const ASYNC_WRITES = true;
 const MAP_SIZE = 1024 * 1024 * 1024 * 10;
 
-async function main() {
+function main() {
   const safeDB = new Lmdb({
     path: './databases/safe/read',
     asyncWrites: ASYNC_WRITES,
@@ -15,9 +15,11 @@ async function main() {
   const value = safeDB.getSync('benchmarkInfo');
   if (!value) throw new Error('Run prepare-read-benchmark.ts');
   const benchmarkInfo = v8.deserialize(value);
+  // eslint-disable-next-line no-console
   console.log(benchmarkInfo);
 
   const {NUM_ENTRIES} = benchmarkInfo;
+  // eslint-disable-next-line no-console
   console.log('(transaction) Reading all entries out');
   safeDB.startReadTransaction();
   {
@@ -30,6 +32,7 @@ async function main() {
     }
     const duration = Date.now() - start;
     const throughput = readEntries.length / duration;
+    // eslint-disable-next-line no-console
     console.log(
       '(transaction) Safe Throughput:',
       throughput,
@@ -38,6 +41,7 @@ async function main() {
   }
   safeDB.commitReadTransaction();
 
+  // eslint-disable-next-line no-console
   console.log('(no-transaction) Reading all entries out');
   {
     const start = Date.now();
@@ -49,6 +53,7 @@ async function main() {
     }
     const duration = Date.now() - start;
     const throughput = readEntries.length / duration;
+    // eslint-disable-next-line no-console
     console.log(
       '(no-transaction) Safe Throughput:',
       throughput,
@@ -57,7 +62,4 @@ async function main() {
   }
 }
 
-main().catch((err) => {
-  console.error(err);
-  process.exitCode = 1;
-});
+main();