+++ packages/core/core/src/Atlaspack.js
@@ -187,6 +187,11 @@ export default class Atlaspack {
         defaultTargetOptions: resolvedOptions.defaultTargetOptions,
         lmdb,
       });
+      if (featureFlags.atlaspackV3CleanShutdown) {
+        this.#disposable.add(() => {
+          rustAtlaspack.end();
+        });
+      }
     }
     this.rustAtlaspack = rustAtlaspack;
 
@@ -378,7 +383,6 @@ export default class Atlaspack {
   |} = {
     /*::...null*/
   }): Promise<BuildEvent> {
-    this.#requestTracker.setSignal(signal);
     let options = nullthrows(this.#resolvedOptions);
     try {
       if (options.shouldProfile) {
@@ -386,6 +390,11 @@ export default class Atlaspack {
       }
       if (options.shouldTrace) {
         tracer.enable();
+        // We need to ensure the tracer is disabled when Atlaspack is disposed as it is a module level object.
+        // While rare (except for tests), if another instance is created later it should not have tracing enabled.
+        this.#disposable.add(() => {
+          tracer.disable();
+        });
       }
       await this.#reporterRunner.report({
         type: 'buildStart',