+++ packages/core/feature-flags/src/index.js
@@ -25,23 +25,21 @@ export const DEFAULT_FEATURE_FLAGS: FeatureFlags = {
   loadableSideEffects: false,
   reduceResolverStringCreation: false,
   inlineBundlesSourceMapFixes: false,
-  conditionalBundlingNestedRuntime: false,
   patchProjectPaths: false,
   cachePerformanceImprovements: process.env.NODE_ENV === 'test',
   environmentDeduplication: false,
   inlineStringReplacementPerf: false,
-  conditionalBundlingAsyncRuntime: false,
-  fixBuildAbortCorruption: false,
   // Default to true as it's a monitoring change. Can be turned off if necessary.
   verboseRequestInvalidationStats: true,
-  conditionalBundlingReporterDuplicateFix: false,
   resolveBundlerConfigFromCwd: false,
-  conditionalBundlingReporterSameConditionFix: false,
-  condbHtmlPackagerChange: false,
-  granularOptionInvalidation: false,
   applyScopeHoistingImprovement: false,
   inlineConstOptimisationFix: false,
   hmrImprovements: false,
+  atlaspackV3CleanShutdown: false,
+  unusedComputedPropertyFix: process.env.NODE_ENV === 'test',
+  emptyFileStarRexportFix: process.env.NODE_ENV === 'test',
+  cliProgressReportingImprovements: false,
+  granularOptionInvalidation: false,
 };
 
 let featureFlagValues: FeatureFlags = {...DEFAULT_FEATURE_FLAGS};