+++ packages/core/integration-tests/test/conditional-bundling.js
@@ -12,7 +12,6 @@ import {
   overlayFS,
   removeDistDirectory,
   runBundles,
-  runBundle,
   distDir,
 } from '@atlaspack/test-utils';
 import sinon from 'sinon';
@@ -79,7 +78,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should have true and false deps as bundles in conditional manifest`, async function () {
-    const dir = path.join(__dirname, 'import-cond-cond-manifest');
+    const workingDir = 'import-cond-cond-manifest';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -91,6 +91,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
 
         index.js:
           const imported = importCond('cond', './a', './b');
@@ -113,7 +124,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -136,7 +149,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should use true bundle when condition is true`, async function () {
-    const dir = path.join(__dirname, 'import-cond-true');
+    const workingDir = 'import-cond-true';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -148,6 +162,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
 
         index.js:
           const conditions = { 'cond': true };
@@ -178,7 +203,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -215,7 +242,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should use both conditional bundles correctly`, async function () {
-    const dir = path.join(__dirname, 'import-cond-both');
+    const workingDir = 'import-cond-both';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -227,6 +255,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
         index.js:
           const conditions = { 'cond1': true, 'cond2': false };
           globalThis.__MCOND = function(key) { return conditions[key]; }
@@ -263,7 +302,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -320,7 +361,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should load false bundle when importing dynamic bundles`, async function () {
-    const dir = path.join(__dirname, 'import-cond-false-dynamic');
+    const workingDir = 'import-cond-false-dynamic';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -332,6 +374,14 @@ describe('conditional bundling', function () {
             "..."
           ]
         }
+      package.json:
+        {
+          "@atlaspack/reporter-conditional-manifest": {
+            "filename": "${path.join(workingDir, 'conditional-manifest.json')}"
+          }
+        }
+
+      yarn.lock: {}
       index.js:
         const conditions = { 'cond': false };
         globalThis.__MCOND = function(key) { return conditions[key]; }
@@ -442,7 +492,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should handle loading conditional bundles when imported in different bundles`, async function () {
-    const dir = path.join(__dirname, 'import-cond-different-bundles');
+    const workingDir = 'import-cond-different-bundles';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -454,6 +505,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
         index.js:
           const conditions = { 'cond1': true, 'cond2': true };
           globalThis.__MCOND = function(key) { return conditions[key]; }
@@ -492,7 +554,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -526,7 +590,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should load bundles in parallel when config enabled`, async function () {
-    const dir = path.join(__dirname, 'import-cond-parallel-enabled');
+    const workingDir = 'import-cond-parallel-enabled';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -545,6 +610,9 @@ describe('conditional bundling', function () {
         {
           "@atlaspack/bundler-default": {
             "loadConditionalBundlesInParallel": true
+          },
+          "@atlaspack/reporter-conditional-manifest": {
+            "filename": "${path.join(workingDir, 'conditional-manifest.json')}"
           }
         }
 
@@ -593,7 +661,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -623,7 +693,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should load conditional bundles in entry html when enabled`, async function () {
-    const dir = path.join(__dirname, 'import-cond-entry-html-enabled');
+    const workingDir = 'import-cond-entry-html-enabled';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -639,6 +710,9 @@ describe('conditional bundling', function () {
         {
           "@atlaspack/packager-html": {
             "evaluateRootConditionalBundles": true
+          },
+          "@atlaspack/reporter-conditional-manifest": {
+            "filename": "${path.join(workingDir, 'conditional-manifest.json')}"
           }
         }
 
@@ -689,7 +763,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -769,7 +845,6 @@ describe('conditional bundling', function () {
       inputFS: overlayFS,
       featureFlags: {
         conditionalBundlingApi: true,
-        condbHtmlPackagerChange: true,
       },
     });
 
@@ -785,7 +860,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should have correct deps as bundles in conditional manifest when nested`, async function () {
-    const dir = path.join(__dirname, 'import-cond-cond-manifest');
+    const workingDir = 'import-cond-cond-manifest-nested';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -797,6 +873,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
 
         index.js:
           const imported = importCond('cond', './a', './b');
@@ -836,7 +923,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );
 
@@ -863,7 +952,8 @@ describe('conditional bundling', function () {
   });
 
   it(`should use load nested bundles when in an async bundle`, async function () {
-    const dir = path.join(__dirname, 'import-cond-false-dynamic');
+    const workingDir = 'import-cond-false-dynamic-nested';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -875,6 +965,14 @@ describe('conditional bundling', function () {
             "..."
           ]
         }
+      package.json:
+        {
+          "@atlaspack/reporter-conditional-manifest": {
+            "filename": "${path.join(workingDir, 'conditional-manifest.json')}"
+          }
+        }
+
+      yarn.lock: {}
       index.js:
         const conditions = { 'cond1': false, 'cond2': true };
         globalThis.__MCOND = function(key) { return conditions[key]; }
@@ -906,7 +1004,6 @@ describe('conditional bundling', function () {
       outputFS: overlayFS,
       featureFlags: {
         conditionalBundlingApi: true,
-        conditionalBundlingNestedRuntime: true,
       },
       defaultConfig: path.join(dir, '.parcelrc'),
     });
@@ -938,89 +1035,9 @@ describe('conditional bundling', function () {
     );
   });
 
-  it(`should support async bundle runtime`, async function () {
-    const dir = path.join(__dirname, 'import-cond-async-bundle-runtime');
-    await overlayFS.mkdirp(dir);
-
-    await fsFixture(overlayFS, dir)`
-      cond-a.js:
-        export default 'cond-a';
-
-      cond-b.js:
-        export default 'cond-b';
-
-      cond-a.html:
-        <script type="module" src="./cond-a.js"></script>
-
-      index.html:
-        <script type="module" src="./index.js"></script>
-
-      index.js:
-        globalThis.__MCOND = (key) => ({ 'cond': true })[key];
-        const condResult = importCond('cond', './cond-a', './cond-b');
-
-        result([condResult]);
-
-      package.json:
-        {
-            "@atlaspack/bundler-default": {
-                "minBundleSize": 0
-            },
-            "@atlaspack/packager-js": {
-              "unstable_asyncBundleRuntime": true
-            },
-            "@atlaspack/packager-html": {
-              "evaluateRootConditionalBundles": true
-            }
-        }
-      yarn.lock:`;
-
-    let b = await bundle(
-      [
-        path.join(__dirname, 'import-cond-async-bundle-runtime/cond-a.html'),
-        path.join(__dirname, 'import-cond-async-bundle-runtime/index.html'),
-      ],
-      {
-        mode: 'production',
-        defaultTargetOptions: {
-          shouldScopeHoist: true,
-          shouldOptimize: false,
-          outputFormat: 'esmodule',
-        },
-        featureFlags: {
-          conditionalBundlingApi: true,
-          conditionalBundlingAsyncRuntime: true,
-        },
-        inputFS: overlayFS,
-      },
-    );
-
-    let indexBundle = b
-      .getBundles()
-      .find((b) => /index.*\.js/.test(b.filePath));
-    if (!indexBundle) return assert.fail();
-
-    let contents = await overlayFS.readFile(indexBundle.filePath, 'utf8');
-    assert(contents.includes('$parcel$global.rwr('));
-
-    let indexHtmlBundle = nullthrows(
-      b.getBundles().find((b) => /index.*\.html/.test(b.filePath)),
-    );
-    let result;
-    await runBundle(b, indexHtmlBundle, {
-      result: (r) => {
-        result = r;
-      },
-    });
-
-    assert.deepEqual(await result, ['cond-a']);
-  });
-
   it(`should have all deps as bundles in conditional manifest when same condition is used multiple times`, async function () {
-    const dir = path.join(
-      __dirname,
-      'import-cond-cond-manifest-same-condition',
-    );
+    const workingDir = 'import-cond-cond-manifest-same-condition';
+    const dir = path.join(__dirname, workingDir);
     await overlayFS.mkdirp(dir);
 
     await fsFixture(overlayFS, dir)`
@@ -1032,6 +1049,17 @@ describe('conditional bundling', function () {
               "..."
             ]
           }
+        package.json:
+          {
+            "@atlaspack/reporter-conditional-manifest": {
+              "filename": "${path.join(
+                workingDir,
+                'conditional-manifest.json',
+              )}"
+            }
+          }
+
+        yarn.lock: {}
 
         index.js:
           const imported1 = importCond('cond', './a', './b');
@@ -1056,7 +1084,6 @@ describe('conditional bundling', function () {
       inputFS: overlayFS,
       featureFlags: {
         conditionalBundlingApi: true,
-        conditionalBundlingReporterSameConditionFix: true,
       },
       defaultConfig: path.join(dir, '.parcelrc'),
     });
@@ -1071,7 +1098,9 @@ describe('conditional bundling', function () {
     // Load the generated manifest
     let conditionalManifest = JSON.parse(
       overlayFS
-        .readFileSync(path.join(distDir, 'conditional-manifest.json'))
+        .readFileSync(
+          path.join(distDir, workingDir, 'conditional-manifest.json'),
+        )
         .toString(),
     );