+++ packages/core/package-manager/test/NodePackageManager.test.js
@@ -12,6 +12,7 @@ import {WORKER_PATH} from '@atlaspack/core';
 import {MockPackageInstaller, NodePackageManager} from '../src';
 
 const FIXTURES_DIR = path.join(__dirname, 'fixtures');
+const ROOT_DIR = path.normalize(path.join(__dirname, '..', '..', '..', '..'));
 
 function normalize(res) {
   return {
@@ -76,6 +77,7 @@ describe('NodePackageManager', function () {
         type: 1,
         invalidateOnFileChange: new Set([
           path.join(FIXTURES_DIR, 'has-foo/node_modules/foo/package.json'),
+          path.join(ROOT_DIR, 'tsconfig.json'),
         ]),
         invalidateOnFileCreate: [
           {
@@ -130,6 +132,7 @@ describe('NodePackageManager', function () {
         type: 1,
         invalidateOnFileChange: new Set([
           path.join(FIXTURES_DIR, 'has-foo/node_modules/a/package.json'),
+          path.join(ROOT_DIR, 'tsconfig.json'),
         ]),
         invalidateOnFileCreate: [
           {
@@ -302,6 +305,7 @@ describe('NodePackageManager', function () {
               FIXTURES_DIR,
               'has-foo/subpackage/node_modules/foo/package.json',
             ),
+            path.join(ROOT_DIR, 'tsconfig.json'),
           ]),
           invalidateOnFileCreate: [
             {