+++ packages/core/test-utils/src/utils.js
@@ -89,14 +89,13 @@ export async function ncp(source: FilePath, destination: FilePath) {
   });
 }
 
-// Mocha is currently run with exit: true because of this issue preventing us
-// from properly ending the workerfarm after the test run:
-// https://github.com/nodejs/node/pull/28788
-//
-// TODO: Remove exit: true in .mocharc.json and instead add the following in this file:
-//   // Spin down the worker farm to stop it from preventing the main process from exiting
-//   await workerFarm.end();
-// when https://github.com/nodejs/node/pull/28788 is resolved.
+after(async () => {
+  // Spin down the worker farm to stop it from preventing the main process from exiting
+  await workerFarm.end();
+  if (isAtlaspackV3) {
+    napiWorkerPool.shutdown();
+  }
+});
 
 const chalk = new _chalk.Instance();
 const warning = chalk.keyword('orange');
@@ -1481,3 +1480,9 @@ it.v3.only = function (...args: mixed[]) {
     origIt.only.apply(this, args);
   }
 };
+
+// If no tests run, then `after()` is not called, and so worker farms are never cleaned up.
+// This test ensures that at least one test runs, and so `after()` is called.
+it('dummy test to ensure there is at least one test', () => {
+  assert(true);
+});