+++ packages/dev/bundle-stats-cli/src/deep-imports.js
@@ -4,7 +4,8 @@ import typeof {loadGraphs} from '@atlaspack/query/src/index';
 import typeof {getBundleStats} from '@atlaspack/reporter-bundle-stats/src/BundleStatsReporter';
 import typeof {PackagedBundle as PackagedBundleClass} from '@atlaspack/core/src/public/Bundle';
 
-module.exports = ((process.env.ATLASPACK_BUILD_ENV === 'production'
+module.exports = ((process.env.ATLASPACK_BUILD_ENV === 'production' ||
+process.env.ATLASPACK_REGISTER_USE_SRC !== 'true'
   ? {
       // Split up require specifier to outsmart packages/dev/babel-register/babel-plugin-module-translate.js
       // $FlowFixMe(unsupported-syntax)