+++ packages/dev/repl/src/atlaspack/BrowserPackageManager.js
@@ -75,12 +75,10 @@ export const BUILTINS = {
 // Package.json fields. Must match package_json.rs.
 const MAIN = 1 << 0;
 const SOURCE = 1 << 2;
-const ENTRIES =
-  MAIN |
-  (process.env.ATLASPACK_BUILD_ENV !== 'production' ||
-  process.env.ATLASPACK_SELF_BUILD
-    ? SOURCE
-    : 0);
+let ENTRIES = MAIN;
+if (process.env.ATLASPACK_REGISTER_USE_SRC === 'true') {
+  ENTRIES |= SOURCE;
+}
 
 export class BrowserPackageManager implements PackageManager {
   resolver: ?ResolverBase;