+++ packages/migrations/parcel-to-atlaspack/src/migrations/migrate-parcelrc.ts
@@ -14,6 +14,7 @@ export type MigrateParcelRcOptions = {
 export async function migrateParcelRc({cwd, dryRun}: MigrateParcelRcOptions) {
   const {default: chalk} = await import('chalk');
 
+  // eslint-disable-next-line no-console
   console.log(chalk.blue('[INFO]'), 'Searching for .parcelrc files');
 
   const parcelrcPaths = await new fdir()
@@ -31,6 +32,7 @@ export async function migrateParcelRc({cwd, dryRun}: MigrateParcelRcOptions) {
       const atlaspckrc = parcelrc.replace(/@parcel\//g, '@atlaspack/');
       if (atlaspckrc !== parcelrc) {
         if (dryRun) {
+          // eslint-disable-next-line no-console
           console.log(
             chalk.blue('[INFO]'),
             `Updated ${relative(cwd, parcelrcPath)}\n${diff(
@@ -50,6 +52,7 @@ export async function migrateParcelRc({cwd, dryRun}: MigrateParcelRcOptions) {
       );
 
       if (dryRun) {
+        // eslint-disable-next-line no-console
         console.log(
           chalk.blue('[INFO]'),
           `Renamed ${relative(cwd, parcelrcPath)}`,
@@ -62,6 +65,7 @@ export async function migrateParcelRc({cwd, dryRun}: MigrateParcelRcOptions) {
     }),
   );
 
+  // eslint-disable-next-line no-console
   console.log(
     chalk.blue('[INFO]'),
     `Migrated ${modifiedFiles.length} .parcelrc files`,