+++ packages/namers/default/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaspack/namer-default
 
+## 2.14.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/plugin@2.14.17
+
+## 2.14.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/plugin@2.14.16
+
 ## 2.14.15
 
 ### Patch Changes