+++ packages/optimizers/image/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @atlaspack/optimizer-image
 
+## 3.1.17
+
+### Patch Changes
+
+- Updated dependencies [[`c75bf55`](https://github.com/atlassian-labs/atlaspack/commit/c75bf553fff4decc285b5fd499a275853b18f8f2)]:
+  - @atlaspack/rust@3.4.0
+  - @atlaspack/utils@2.16.1
+  - @atlaspack/workers@2.14.17
+  - @atlaspack/plugin@2.14.17
+
+## 3.1.16
+
+### Patch Changes
+
+- Updated dependencies [[`30ee2cf`](https://github.com/atlassian-labs/atlaspack/commit/30ee2cfcd34cf2646ded0eda13fdb80a2a5de529)]:
+  - @atlaspack/utils@2.16.0
+  - @atlaspack/workers@2.14.16
+  - @atlaspack/plugin@2.14.16
+
 ## 3.1.15
 
 ### Patch Changes